==> Capstone_Submission/1_compiling_IMDB_data.ipynb <==
import requests
import pandas as pd
import time
import json

#https://www.imdb.com/interfaces/

There is a significant amount of data provided by IMDB. In this notebook, I parse through all of it and merge the relevant information into a single csv.

a = pd.read_csv('./data/title.akas.tsv.gz', sep='\t', compression='gzip')
b = pd.read_csv('./data/title.basics.tsv.gz', sep='\t', compression='gzip')
c = pd.read_csv('./data/title.crew.tsv.gz', sep='\t', compression='gzip')
d = pd.read_csv('./data/title.episode.tsv.gz', sep='\t', compression='gzip')
e = pd.read_csv('./data/title.principals.tsv.gz', sep='\t', compression='gzip')
f = pd.read_csv('./data/title.ratings.tsv.gz', sep='\t', compression='gzip')
g = pd.read_csv('./data/name.basics.tsv.gz', sep='\t', compression='gzip')

#renaming the unique id for consistency
a.rename(columns = {'titleId':'tconst'}, inplace=True)

a.shape

d.shape

#merging dataframes a and d
a = a.merge(d, on='tconst', how='left')

a.shape

a.head()

#nulls are coded as '\N' so we need a function to replace these nulls
def slash_n(x):
    if x == '\\N':
        return None
    else:
        return x

#filling the nulls
a['episodeNumber'] = a['episodeNumber'].apply(slash_n)

#if an entry doesn't contain an episode number, then it must be a movie. 
a = a[a['episodeNumber'].isnull() == True]

#overwriting dataframe d to save memory
d = 'x'

temp = a[a['language'] == '\\N']

#restricting the dataset to english movies
a = a[a['language'] == 'en']

backup = a

# #loading in more data
# b = pd.read_csv('./data/title.basics.tsv.gz', sep='\t', compression='gzip')

b.shape

#merging the new data onto our other dataset
a = a.merge(b, on='tconst', how='left')

a.shape

#overwriting b to save memory
b='x'

backup = a

# #reading in more data
# c = pd.read_csv('./data/title.crew.tsv.gz', sep='\t', compression='gzip')

#merging onto existing data
a = a.merge(c, on='tconst', how='left')

#overwriting c to save memory
c = 'x'

a.shape

backup = a

#reading in more data
# f = pd.read_csv('./data/title.ratings.tsv.gz', sep='\t', compression='gzip')
# g = pd.read_csv('./data/name.basics.tsv.gz', sep='\t', compression='gzip')

#merging onto existing data
a = backup.merge(f, on='tconst', how='left')

a.head()

#dropping unnecessary columns
a.drop(columns=['seasonNumber',
                'episodeNumber', 
                'parentTconst', 
                'isAdult',  
                'language', 
                'titleType',
                'types',
                'attributes'], inplace=True)

a.columns

#getting rid of duplicate entries for films
a = a.drop_duplicates(subset=['tconst'], keep = 'first')

#filling nulls
a['endYear'] = a['endYear'].apply(slash_n)

#if a title has an end year, it is a tv series
a = a[a['endYear'].isnull() == True]

#reading in and merging more data
# e = pd.read_csv('./data/title.principals.tsv.gz', sep='\t', compression='gzip')
a = a.merge(e, on='tconst', how='left')

temp_list = []
for unique_id in a['tconst']:
    temp_list.append(unique_id)

e_temp = e[e['tconst'].isin(temp_list) == True]

e_temp.shape

e is a massive dataframe providing details on the cast and crew of a movie, so in the following cells I break it down so that it ran better on my computer. 

#compiling all the people that worked on each movie into a single cell
e_temp_1 = e_temp[:500000]
e_temp_1 = e_temp_1.groupby('tconst').apply(lambda x: x.sum())
e_temp_1.drop(columns=['ordering', 'category', 'job', 'characters'], inplace=True)
e_temp_1['nconst'] = e_temp_1['nconst'].apply(lambda x: x.split('nm')).apply(lambda x: x[1:])

#compiling all the people that worked on each movie into a single cell
e_temp_2 = e_temp[500000:1000000]
e_temp_2 = e_temp_2.groupby('tconst').apply(lambda x: x.sum())
e_temp_2.drop(columns=['ordering', 'category', 'job', 'characters'], inplace=True)
e_temp_2['nconst'] = e_temp_2['nconst'].apply(lambda x: x.split('nm')).apply(lambda x: x[1:])

#compiling all the people that worked on each movie into a single cell
e_temp_3 = e_temp[1000000:]
e_temp_3 = e_temp_3.groupby('tconst').apply(lambda x: x.sum())
e_temp_3.drop(columns=['ordering', 'category', 'job', 'characters'], inplace=True)
e_temp_3['nconst'] = e_temp_3['nconst'].apply(lambda x: x.split('nm')).apply(lambda x: x[1:])

cast_crew = pd.concat([e_temp_1, e_temp_2, e_temp_3])
cast_crew.drop(columns=['tconst'], inplace=True)
cast_crew.reset_index(inplace=True)

e_temp = e_temp.groupby('tconst').apply(lambda x: x.sum())
e_temp.drop(columns=['ordering', 'job', 'characters'], inplace=True)
e_temp['nconst'] = e_temp['nconst'].apply(lambda x: x.split('nm')).apply(lambda x: x[1:])

# part_1 = e_temp[:500000]

# temp_df = pd.DataFrame(columns = e_temp.category.unique())
# temp_df['tconst'] = []
# temp_df.set_index('tconst', inplace=True)

# for index, row in part_1.iterrows():
#     temp_df.loc[row['tconst'], row['category']] = row['nconst']

# for index, row in part_1.iterrows():
#     temp_df.loc[row['tconst'], row['category']] = row['nconst']

# part_2 = e_temp[500000:750000]

# temp_df2 = pd.DataFrame(columns = e_temp.category.unique())
# temp_df2['tconst'] = []
# temp_df2.set_index('tconst', inplace=True)

# for index, row in part_2.iterrows():
#     temp_df2.loc[row['tconst'], row['category']] = row['nconst']

# part_3 = e_temp[750000:1000000]

# temp_df3 = pd.DataFrame(columns = e_temp.category.unique())
# temp_df3['tconst'] = []
# temp_df3.set_index('tconst', inplace=True)

# for index, row in part_3.iterrows():
#     temp_df3.loc[row['tconst'], row['category']] = row['nconst']

# part_4 = e_temp[1000000:1250000]

# temp_df4 = pd.DataFrame(columns = e_temp.category.unique())
# temp_df4['tconst'] = []
# temp_df4.set_index('tconst', inplace=True)

# for index, row in part_4.iterrows():
#     temp_df4.loc[row['tconst'], row['category']] = row['nconst']

# part_5 = e_temp[1250000:]

# temp_df5 = pd.DataFrame(columns = e_temp.category.unique())
# temp_df5['tconst'] = []
# temp_df5.set_index('tconst', inplace=True)

# for index, row in part_5.iterrows():
#     temp_df5.loc[row['tconst'], row['category']] = row['nconst']

# people_info = pd.concat([temp_df, temp_df2, temp_df3, temp_df4, temp_df5])
# people_info.reset_index(inplace=True)
# people_info = people_info.drop_duplicates(subset=['tconst'], keep='first')

final_df = a.merge(cast_crew, on='tconst', how='left')

final_df.shape

final_df.to_csv('./final.csv')

# final_df.to_csv('./data/final.csv', index=False)

# test = e_temp[:1000]
# test = test.groupby('tconst').apply(lambda x: x.sum())
# test.drop(columns=['tconst', 'ordering', 'category', 'job', 'characters'], inplace=True)
# test['nconst'] = test['nconst'].apply(lambda x: x.split('nm')).apply(lambda x: x[1:])
