==> Capstone_Submission/3_wikipedia_synopsis_scraping.ipynb <==
This notebook took slightly over 2 hours to run on my computer, so I do not recommend running it. 

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

import requests

import wikipediaapi

movies = pd.read_csv('./data/for_wiki_scraping.csv')
movies.drop(columns=['Unnamed: 0', 'Unnamed: 0.1'], inplace=True)

#creating a function to grab the sections from a movie's wiki page
def get_wiki_info(page):
    temp_dict = {}
    for header in page.sections:
        temp_dict[header.title] = header.text
    return temp_dict

temp_list = []
wiki_wiki = wikipediaapi.Wikipedia('en')

for movie_title in movies['wiki_title']:
    try:
        #creating a blank dictionary
        temp_dict = {}
        #checking if page exists
        page = wiki_wiki.page(movie_title.replace(' ', '_'))
        page2 = wiki_wiki.page((movie_title[:-12] + ' (film)').replace(' ', '_'))
        page3 = wiki_wiki.page((movie_title[:-12]).replace(' ', '_'))

        if page.exists() == True:
            temp_list.append(get_wiki_info(page))
        elif page2.exists() == True:
                temp_list.append(get_wiki_info(page2))
        elif page3.exists() == True:
                temp_list.append(get_wiki_info(page3))
        else:
            temp_list.append({'plot':'ERROR'})
    except:
        temp_list.append({'plot':'ERROR'})
        
movies['wiki_scrape'] = temp_list
movies.to_csv('./data/with_wiki_scrape_complete.csv')
