==> Capstone_Submission/4_compiling_wikipedia.ipynb <==
import pandas as pd
import numpy as np
import ast
import matplotlib.pyplot as plt

import spacy
from spacytextblob.spacytextblob import SpacyTextBlob

import re

movies = pd.read_csv('./data/with_wiki_scrape_complete.csv')
movies.drop(columns=['Unnamed: 0'], inplace=True)

movies.head()

#creating a list of all of the wiki info converted to actual dictionaries
dict_list = []

for i in range(0, 10000):
    
    dict_list.append(ast.literal_eval(movies.iloc[i, 17]))
    
#creating a list of the plot info, filling in "ERROR" if plot not called plot
plot_list = []

for dictionary in dict_list:
    try: 
        plot_list.append(dictionary['Plot'])
    except:
        try: plot_list.append(dictionary['Summary'])
        except:
            try:
                plot_list.append(dictionary['Synopsis'])
            except:
                plot_list.append('ERROR')
        
movies['plot'] = plot_list

movies[movies['plot'] == "ERROR"].shape

we have 515 movies that we were unable to scrape a plot for. Luckily we have some back up short plot summaries that we can use. 

#selecting just the plot synopsis from our "scraped_data" column
movies['backup_plot'] = movies['scraped_data'].apply(
    lambda x: x.split("\"Plot\":")[1].split('\"Language\":')[0][:-2])

movies.columns

#replacing the movies with missing plots with the backup plots
#grouping the movies that don't have "ERROR" as their plot
movies_long_plot = movies[movies['plot'] != 'ERROR']
#excluding the blank plots
movies_long_plot = movies_long_plot[movies_long_plot['plot'] != '']
#selecting error plots
movies_plot_error = movies[movies['plot'] == 'ERROR']
movies_plot_error['plot'] = movies_plot_error['backup_plot']
#selecting blank plots
movies_no_plot = movies[movies['plot'] == '']
movies_no_plot['plot'] = movies_no_plot['backup_plot']
#putting it all back together
movies = pd.concat([movies_long_plot, movies_plot_error, movies_no_plot])
movies.sort_values('votes', ascending=False, inplace=True)

movies.columns

movies.shape

#making the "genres" column back into a list
movies['genres'] = movies['genres'].apply(lambda x: x.replace(' ', '')\
    .replace('\'', '').replace('[', '').replace(']','').split(','))

#separating the genres into individual columns
genre_1 = []
genre_2 = []
genre_3 = []

for genre in movies['genres']:
    genre_1.append(genre[0])

for genre in movies['genres']:
    try:
        genre_2.append(genre[1])
    except:
        genre_2.append(np.nan)

for genre in movies['genres']:
    try:
        genre_3.append(genre[2])
    except:
        genre_3.append(np.nan)
#adding the lists as columns        
movies['genre_1'] = genre_1
movies['genre_2'] = genre_2
movies['genre_3'] = genre_3


#making the directors column into a list
movies['directors'] = movies['directors'].apply(lambda x: x.split(','))

# separating directors into distinct columns
director_1 = []
director_2 = []
director_3 = []

for director in movies['directors']:
    director_1.append(director[0])

for director in movies['directors']:
    try:
        director_2.append(director[1])
    except:
        director_2.append(np.nan)

for director in movies['directors']:
    try:
        director_3.append(director[2])
    except:
        director_3.append(np.nan)
#adding the lists as columns        
movies['director_1'] = genre_1
movies['director_2'] = genre_2
movies['director_3'] = genre_3


# making the writers column into a list
movies['writers'] = movies['writers'].apply(lambda x: x.split(','))

# separating writers into distinct columns
writer_1 = []
writer_2 = []
writer_3 = []

for writer in movies['writers']:
    writer_1.append(writer[0])

for writer in movies['writers']:
    try:
        writer_2.append(writer[1])
    except:
        writer_2.append(np.nan)

for writer in movies['writers']:
    try:
        writer_3.append(writer[2])
    except:
        writer_3.append(np.nan)
#adding the lists as columns        
movies['writer_1'] = writer_1
movies['writer_2'] = writer_2
movies['writer_3'] = writer_3

# making the cast/crew into a list
movies['cast_crew'] = movies['cast_crew'].apply(lambda x: x.replace(' ', '')\
    .replace('\'', '').replace('[', '').replace(']','').split(','))

movies['cast_crew'].head()

# separating cast_crew into distinct columns using a function

def movie_people(position):
    cast_list = []
    for person_list in movies['cast_crew']:
        try:
            cast_list.append(person_list[position])
        except:
            cast_list.append(np.nan)
    return cast_list
        
# calling the function to create the columns        
movies['cast_1'] = movie_people(0)
movies['cast_2'] = movie_people(1)
movies['cast_3'] = movie_people(2)
movies['cast_4'] = movie_people(3)
movies['cast_5'] = movie_people(4)
movies['cast_6'] = movie_people(5)
movies['cast_7'] = movie_people(6)
movies['cast_8'] = movie_people(7)
movies['cast_9'] = movie_people(8)
movies['cast_10'] = movie_people(9)

#creating a decade tag 
movies['decade'] = movies['release_year'].apply(lambda x: round((x-4), -1))

movies.head()

movies.columns

movies.describe()

#dropping columns that we will not be using further
movies.drop(columns = [
    'wiki_scrape',
    'backup_plot',
    'scraped_data',
    'title',
    'original_title',
    'wiki_title',
    'directors',
    'writers',
    'genres',
    'cast_crew',
    'decade'
], inplace=True)

#sentiment analysis

#removing unnecessary characters
movies['plot'] = movies['plot'].apply(
    lambda x: str(x.replace('\n', '').replace("\\", '')))

movies['plot'] = [re.sub('[\(\[].*?[\)\]]', '', t.replace('\n', '')) for t in movies['plot']]

movies['plot'] = [t.replace('/\\', ' ').replace(':-)', ' ').replace('tdb> ', ' ') 
              for t in movies['plot']]

movies['plot'] = [t.replace('=', ' ').replace('--', ' ').replace('_', ' ')\
              .replace('}', ' ').replace('*', ' ').replace('^', ' ')\
              .replace('~', ' ')
              for t in movies['plot']]

movies['plot'] = [t.replace('- < > -', ' ').replace('|', ' ').replace('*-', ' ') 
              for t in movies['plot']]

movies['plot'] = [t.replace('...', ' ').replace('\t', ' ').lower().strip() 
              for t in movies['plot']]



temp = movies[:50]

#creating a column that indicates the polarity of the synopsis, hopefully to
#reveal if a movie is a feel good movie or not

#setting up sentiment analysis and creating column

nlp = spacy.load('en_core_web_sm')

nlp.add_pipe('spacytextblob')

movies['plot_polarity'] = movies['plot'].apply(lambda x: nlp(x)._.polarity)  

movies.to_csv('./data/data_for_eda.csv')
