==> Capstone_Submission/6_preparing_review_data.ipynb <==
import pandas as pd
import numpy as np
from thefuzz import process
import imdb
import fuzzyset

films = pd.read_csv('./data/ml-25m/ml-25m/movies.csv')
films.drop(columns=['genres'], inplace=True)

ratings = pd.read_csv('./data/ml-25m/ml-25m/ratings.csv')
ratings.drop(columns=['timestamp'], inplace=True)

films.head()

#removing the release year and making it its own column
films['release_year'] =  films['title'].apply(lambda x: x[-5:-1])

# removing parentheticals from titles for searching purposes:
films['primary_title'] = films['title'].apply(lambda x: x.split('(')[0][:-1])

#fixing titles that have "the" moved to the end for matching purposes
test_list = []
for title in films['primary_title']:
    if str(title[-3:]) == 'The':
        test_list.append('The ' + str(title[:-5]))
    else:
        test_list.append(title)
films['primary_title'] = test_list

films.head()

movies = pd.read_csv('./data/data_for_eda.csv')

movies.head()

#for films, adding the year back into the title to help with matching efficiency
match_title = []
title_list = films['primary_title']
year_list = films['release_year']
for i in range(0,len(title_list)):
    match_title.append(title_list[i] + ' (' + year_list[i] + ')')
    
films['match_title'] = match_title

#for movies, adding the year back into the title to help with matching efficiency
match_title = []
title_list = movies['primary_title']
year_list = movies['release_year']
for i in range(0,len(title_list)):
    match_title.append(title_list[i] + ' (' + str(year_list[i]) + ')')
    
movies['match_title'] = match_title

movies.head()

films.head()

#finding out which films need to be fuzzy joined, since there are some direct matches
movies = pd.merge(movies, films, on='primary_title', how='left')
movies.isnull().sum()
#there are 2,125 that do not have a direct title match. we will subset accordingly

# movies.drop(columns=['release_year_y', 'match_title_y'], inplace=True)

movies.dtypes

movies['release_year_x'] = movies['release_year_x'].apply(lambda x: str(x))

movies.head()

movies_fuzzy = movies[movies['movieId'].isnull() == True]

films.head()

title_list = fuzzyset.FuzzySet()
for title in films['match_title']:
    title_list.add(title)
    
match_list = []
for film_title in movies['primary_title']:
    match_list.append(title_list.get(film_title)[0][1])
    
movies['match_list'] = match_list

movies.head()

#getting a list of the films that have been properly matched

year_list_1 = list(movies['match_title_x'])
year_list_2 = list(movies['match_list'])
titles = list(movies['primary_title'])
matched = []

for i in range(0, len(year_list_1)):
    if year_list_1[i][-6:] == year_list_2[i][-6:]:
        matched.append(titles[i])

#reducing the dataset to just movies that have been paired properly
movies = movies[movies['primary_title'].isin(matched)==True]

#we may have some duplicates but can deal with those later. 

#subsetting to keep only movies where the match_list title year matches the release year

prim_title_year = list(movies['release_year_x'])
match_list_year = []
for title in movies['match_list']:
    match_list_year.append(title[-5:-1])

#we have 9091 titles that have reviews that we can match them to. 
len(matched)

matched_movies = []
primary_title = list(movies['primary_title'])
for i in range(0, len(primary_title)):
    if prim_title_year[i] == match_list_year[i]:
        matched_movies.append(primary_title[i])

movies = movies[movies['primary_title'].isin(matched_movies) == True]

#subsetting to keep only movies where the match_list title actually matches the primary_title
prim_title_year = list(movies['release_year_x'])
match_list_year = []
for title in movies['match_list']:
    match_list_year.append(title[-5:-1])
matched_movies = []

primary_title = list(movies['primary_title'])
for i in range(0, len(primary_title)):
    if prim_title_year[i] == match_list_year[i]:
        matched_movies.append(primary_title[i])

movies = movies[movies['primary_title'].isin(matched_movies)==True]

movies.drop_duplicates(subset='match_title_x', inplace=True)

#getting just the matching title and movie ID so we can add the ID to our movies dataset. 
films2 = pd.read_csv('./data/ml-25m/ml-25m/movies.csv')

#removing the release year and making it its own column
films2['release_year'] =  films2['title'].apply(lambda x: x[-5:-1])

# removing parentheticals from titles for searching purposes:
films2['primary_title'] = films2['title'].apply(lambda x: x.split('(')[0][:-1])

#fixing titles that have "the" moved to the end for matching purposes
test_list = []
for title in films2['primary_title']:
    if str(title[-3:]) == 'The':
        test_list.append('The ' + str(title[:-5]))
    else:
        test_list.append(title)
films2['primary_title'] = test_list

#for films, adding the year back into the title to help with matching efficiency
match_title = []
title_list = films2['primary_title']
year_list = films2['release_year']
for i in range(0,len(title_list)):
    match_title.append(title_list[i] + ' (' + year_list[i] + ')')
    
films2['match_title'] = match_title

films2.drop(columns=['title', 'genres', 'release_year'], inplace=True) 

films2.rename(columns={'match_title':'match_list'}, inplace=True)

movies.drop(columns=['movieId', 'title', 'release_year_y', 'match_title_y', 'match_title_x'], inplace=True)

movies = pd.merge(movies, films2, how='left', on='match_list')

movies.drop(columns=['match_list', 'primary_title_y', 'match_list'], inplace=True)

movies.to_csv('./data/movies_with_review_id.csv', index=False)
